qcacmn: Change the datatype of resp_id
The datatype of resp_id in
wmi_afc_power_event_param is 32 bit.

Change the datatype of resp_id from
uint8_t to uint32_t
in the structure reg_fw_afc_power_event,
to write
QCA_WLAN_VENDOR_ATTR_AFC_POWER_EVT_RESP_ID
in the
QCA_WLAN_VENDOR_AFC_POWER_UPDATE_COMPLETE_EVENT

CRs-Fixed: 3159542

+++ umac/regulatory/dispatcher/inc/reg_services_public_struct.h
@@ -1361,7 +1361,7 @@ struct reg_afc_expiry_event {
  * @afc_chan_info: Pointer to AFC channel object
  */
 struct reg_fw_afc_power_event {
-	uint8_t resp_id;
+	uint32_t resp_id;
 	enum reg_fw_afc_power_event_status_code fw_status_code;
 	enum reg_afc_serv_resp_code serv_resp_code;
 	uint32_t afc_wfa_version;
